test: add delete-highway-by-id use-case tests

+++ src/application/use-cases/delete-highway-by-id.ts
@@ -14,6 +14,12 @@ export class DeleteHighwayById {
       throw new Error('`id` should be number')
     }
 
+    const highway = await this.highwaysRepository.findById(id)
+
+    if (!highway) {
+      throw new Error('highway not found')
+    }
+
     await this.highwaysRepository.deleteById(id)
   }
 }
